Refactoring: certificates and lifecycle
- a function related to certificates moved to a proper package
- lifecycle container func now accepts binary path as param
